refactor: extract helpers/gateway.py (Phase 6.5)
Moves the OpenClaw gateway WebSocket RPC client + HTTP /tools/invoke
fan-out out of dashboard.py into helpers/gateway.py — following the same
pattern as helpers/streams.py (Phase 6.3). dashboard.py now 24,956
lines (was 25,121; −165), with a 212-line self-contained module
owning `_ws_client` / `_ws_lock` / `_ws_connected`, `_gw_ws_connect`,
`_gw_ws_rpc`, `_gw_invoke`, and `_gw_invoke_docker`.

Late `import dashboard as _d` (the same trick routes/*.py already use)
lets the helper reach `_load_gw_config`, `__version__`,
`_CURRENT_PLATFORM`, `_uuid`, and `_urllib_req` without creating a
circular import. Only CLI/config-intertwined pieces
(`_load_gw_config`, `_detect_gateway_token`, `_detect_gateway_port`,
`_auto_discover_gateway`, `GATEWAY_URL/TOKEN` globals) stay in
dashboard.py.

Re-exports `_gw_invoke`, `_gw_invoke_docker`, `_gw_ws_rpc` from
dashboard.py so existing `_d.<name>` callers in routes/crons.py,
routes/health.py, routes/meta.py, routes/overview.py, and
routes/sessions.py keep working unchanged.

Smoke tested against a live local gateway: `/api/sessions` returns 1
session via gateway RPC, `/api/crons` returns 1, `/api/overview`
returns model=@oi/beta, and all 35 probed endpoints return 200 with no
tracebacks.
